adding helper in configs
