+++ starter_code/configs/hbs.configs.js
@@ -1,0 +1,6 @@
+const hbs = require('hbs');
+const path = require('path');
+
+hbs.registerPartials(path.join(__dirname, '../views/partials'));
+
+require('../helpers/split.helper');
